@@ -1,9 +1,15 @@
 module github.com/elastic/gosigar
 
-go 1.14
+go 1.22.9
 
 require (
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.4.0
-	golang.org/x/sys v0.0.0-20180810173357-98c5dad5d1a0
+	golang.org/x/sys v0.26.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
