@@ -15,7 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//+build ignore
+//go:build ignore
+// +build ignore
 
 package main
 
